Added read files exist check

+++ scripts/spades_hybrid.sh
@@ -5,9 +5,30 @@ reads2=$ofolder/data/fastq/miseq2.fastq
 long_reads=$ofolder/data/fastq/nanopore.fastq
 
 
-mkdir -p results/spades
-cd results/spades
-$ofolder/src/spades/bin/spades.py --pe1-1 $reads1 --pe1-2 $reads2 --nanopore $long_reads -t 15 -o miseq_ont
 
-mv contigs.fasta hybridspades_contigs.fasta
-mv scaffolds.fasta hybridspades_scaffolds.fasta
+if [ -f $reads1 ]; then check1=`head -1 $reads1`; fi
+if [ -f $reads2 ]; then check2=`head -1 $reads2`; fi
+if [ -f $long_reads ]; then check3=`head -1 $long_reads`; fi
+
+
+if [ ! -f $reads1 ]; then
+        echo; echo "  Could not find read-file "  ${reads1}; echo
+elif  [ ! -f $reads2 ]; then
+        echo; echo "  Could not find read-file "  ${reads2}; echo
+elif  [ ! -f $long_reads ]; then
+        echo; echo "  Could not find read-file "  ${long_reads}; echo
+elif [ -z "${check1}" ]; then
+        echo; echo "  The read-file "  ${reads1} is empty!; echo
+elif [ -z "${check2}" ]; then
+        echo; echo "  The read-file "  ${reads2} is empty!; echo
+elif [ -z "${check3}" ]; then
+        echo; echo "  The read-file "  ${long_reads} is empty!; echo
+else
+
+	mkdir -p results/spades
+	cd results/spades
+	$ofolder/src/spades/bin/spades.py --pe1-1 $reads1 --pe1-2 $reads2 --nanopore $long_reads -t 15 -o miseq_ont
+
+	mv contigs.fasta hybridspades_contigs.fasta
+	mv scaffolds.fasta hybridspades_scaffolds.fasta
+fi
